refactor: check if the domain alias and resource class match
