+++ models/classes/import/ServiceProvider/ImportServiceProvider.php
@@ -25,8 +25,8 @@ namespace oat\tao\model\import\ServiceProvider;
 use oat\tao\model\upload\UploadService;
 use oat\tao\model\import\service\AgnosticImportHandler;
 use oat\generis\model\DependencyInjection\ContainerServiceProviderInterface;
+use oat\tao\model\StatisticalMetadata\Import\ImportStatisticalMetadataProcessor;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
-use oat\tao\model\StatisticalMetadata\Import\Processor\ImportStatisticalMetadataProcessor;
 
 use function Symfony\Component\DependencyInjection\Loader\Configurator\service;
